feat: expose stop() from launcher.connection and document programmatic use
Move the click-independent stop() implementation (and its private helpers)
from ansys.mapdl.core.cli.stop into ansys.mapdl.core.launcher.connection, so
close_all_local_instances() no longer needs to import from the cli package
and stop() can be reused from regular Python scripts without needing to
launch_mapdl() first. cli/stop.py now re-exports stop() and keeps only the
Click wrapper. stop is also importable as ansys.mapdl.core.launcher.stop.

Benchmarked with hyperfine (30 runs) to confirm the move does not regress
`pymapdl stop` CLI runtime: 247.9 ms +/- 13.3 ms before vs 252.4 ms +/- 17.9 ms
after, a difference within run-to-run noise.

Add user guide documentation showing how to stop MAPDL instances both from
the CLI and programmatically with stop() and close_all_local_instances(), and
update the API reference to point at the new canonical location.

Closes #4297
