+++ tests/test_cli.py
@@ -147,7 +147,7 @@ def test_pymapdl_stop_instances(run_cli, mapping):
     fake_processes = [make_fake_process(**each) for each in fake_process_]
 
     with (
-        patch("ansys.mapdl.core.cli.stop._kill_process") as mock_kill,
+        patch("ansys.mapdl.core.launcher.connection._kill_process") as mock_kill,
         patch("psutil.pid_exists") as mock_pid,
         patch("psutil.process_iter", return_value=iter(fake_processes)),
     ):
@@ -272,7 +272,7 @@ def test_pymapdl_stop_permission_handling(run_cli):
         patch("psutil.process_iter", return_value=test_processes),
         patch("psutil.pid_exists", return_value=True),
         patch(
-            "ansys.mapdl.core.cli.stop._kill_process",
+            "ansys.mapdl.core.launcher.connection._kill_process",
             side_effect=mock_kill_process,
         ),
     ):
@@ -331,7 +331,7 @@ def test_pymapdl_stop_with_username_containing_domain(run_cli):
         patch("psutil.process_iter", return_value=[mock_process]),
         patch("psutil.pid_exists", return_value=True),
         patch(
-            "ansys.mapdl.core.cli.stop._kill_process",
+            "ansys.mapdl.core.launcher.connection._kill_process",
             side_effect=mock_kill_process,
         ),
     ):
@@ -1982,7 +1982,7 @@ def test_stop_all_kill_raises_no_such_process(run_cli):
         patch("psutil.process_iter", return_value=[proc]),
         patch("psutil.pid_exists", return_value=True),
         patch(
-            "ansys.mapdl.core.cli.stop._kill_process",
+            "ansys.mapdl.core.launcher.connection._kill_process",
             side_effect=psutil.NoSuchProcess(12345),
         ),
     ):
@@ -2005,7 +2005,7 @@ def test_stop_all_process_raises_during_iteration(run_cli):
     with (
         patch("psutil.process_iter", return_value=[proc]),
         patch("psutil.pid_exists", return_value=True),
-        patch("ansys.mapdl.core.cli.stop._kill_process"),
+        patch("ansys.mapdl.core.launcher.connection._kill_process"),
     ):
         output = run_cli("stop --all")
     assert "error: no ansys instances have been found." in output.lower()
@@ -2022,7 +2022,7 @@ def test_stop_port_kill_raises(run_cli):
         patch("psutil.process_iter", return_value=[fake_proc]),
         patch("psutil.pid_exists", return_value=True),
         patch(
-            "ansys.mapdl.core.cli.stop._kill_process",
+            "ansys.mapdl.core.launcher.connection._kill_process",
             side_effect=psutil.NoSuchProcess(12345),
         ),
     ):
@@ -2072,7 +2072,9 @@ def test_stop_pid_kills_children(run_cli):
 
     with (
         patch("psutil.Process", return_value=main_mock),
-        patch("ansys.mapdl.core.cli.stop._kill_process", side_effect=track_kill),
+        patch(
+            "ansys.mapdl.core.launcher.connection._kill_process", side_effect=track_kill
+        ),
     ):
         output = run_cli(f"stop --pid {pid}")
 
@@ -2092,7 +2094,7 @@ def test_stop_pid_kill_failed(run_cli):
 
     with (
         patch("psutil.Process", return_value=mock_proc),
-        patch("ansys.mapdl.core.cli.stop._kill_process"),
+        patch("ansys.mapdl.core.launcher.connection._kill_process"),
     ):
         output = run_cli(f"stop --pid {pid}")
     assert "could not be killed" in output.lower()
@@ -2100,7 +2102,7 @@ def test_stop_pid_kill_failed(run_cli):
 
 def test_kill_process():
     """Test _kill_process calls proc.kill() (line 175)."""
-    from ansys.mapdl.core.cli.stop import _kill_process
+    from ansys.mapdl.core.launcher.connection import _kill_process
 
     mock_proc = MagicMock()
     _kill_process(mock_proc)
